family-studies - Improve support for family studies Moved Pedigree to org.phenotips.studies.family package.
