+++ components/family-studies/api/src/main/java/org/phenotips/studies/family/script/FamilyScriptService.java
@@ -21,9 +21,9 @@ import org.phenotips.data.Patient;
 import org.phenotips.data.PatientRepository;
 import org.phenotips.studies.family.Family;
 import org.phenotips.studies.family.FamilyRepository;
+import org.phenotips.studies.family.Pedigree;
 import org.phenotips.studies.family.Processing;
 import org.phenotips.studies.family.internal.JSONResponse;
-import org.phenotips.studies.family.internal.Pedigree;
 import org.phenotips.studies.family.internal.StatusResponse;
 import org.phenotips.studies.family.internal.export.XWikiFamilyExport;
